moved `qemu` module
The `qemu` module was moved from `library/helper/tests`
toml`library/helper/miscellaneous`.
