+++ kernel/src/library/helper/test.rs
@@ -75,43 +75,3 @@ fn trivial_assertion()
 	assert_eq!(1, ONE);
 	assert_eq!(ONE, 1);
 }
-
-/// ## QEMU Abstractions
-///
-/// Contains helpers for running the kernel with QEMU.
-pub mod qemu
-{
-
-	/// ### Write An Exit Code
-	///
-	/// Writes to the `0xF4` port the correct bytes that indicate
-	/// either success or failure.
-	#[inline]
-	fn exit(success: bool)
-	{
-		use qemu_exit::QEMUExit;
-
-		#[cfg(target_arch = "x86_64")]
-		let qemu_exit_handle = qemu_exit::X86::new(0xF4, 0x3);
-
-		if success {
-			qemu_exit_handle.exit_success();
-		} else {
-			qemu_exit_handle.exit_failure();
-		}
-	}
-
-	/// ### Exit QEMU With Success
-	///
-	/// Write a success exit code for QEMU to recognize and exit.
-	#[allow(dead_code)]
-	#[inline]
-	pub fn exit_with_success() { exit(true); }
-
-	/// ### Exit QEMU Without Success
-	///
-	/// Write a failure exit code for QEMU to recognize and exit.
-	#[allow(dead_code)]
-	#[inline]
-	pub fn exit_with_failure() { exit(false); }
-}
